Creating admin analytics page to make the analytics menu section in sidebar functionable
Making the deisgn same with the view elections
by presenting the election cards and its filtering
options.

Adding analytics.php in each specific elections
to handle the analytic page for each selected
elections that shows the report of vote turnouts
no. candidates and registered voters, winner
candidates, and graphs and table.

Used chart.js for the dynamic graph that fetches
specific data in db based on the scope of the
admin, election, and voters.

Fixing bugs and errors in fetching data

Fixing bugs in deletion of admin super admin
manage admin page.