Modified the whole admin_dashboard and admin analytics pages to centralized data fetching helpers with the use of analytics_scopes, election_helpers, and auth helpers
added new visuals for analytics section (charts and graphs)
Added the inheritance of previous admin to the new admin to
pass or inherit it scopes (all electiopns, users, analytics,
and candidates management)

Making all election that only one admin is allowed to manage
by status categorization

added the Others-Default and Non-Academic-Student admin admin_dashboard
and analytics pages per election

addded range of years comparison from the selected range
of years to the current or selected years

added comparison of election if there's an election in the
same year

modified the creation of election to make the COOP selection to
Others so super admin can choose if the Others election
is for COOP or default others by toggling the MIGS
restrictions

Modified the system flow to be aligned in the system revisions

Added the Abstained computation in the analytics

Modified the voters dashboard to be aligned in the analytics
scopes helpers

modifed the manage candidates, view elections, view vote counts,
add and edit candidates, manage users to matched with the inheritance
and new admin scopes

for the reports generation, I used TCPDF Libraries

make it aligned in the analytics_scopes, admin_functions helpers
so the displaying of data is aligned with the sql and analytics.

Modified some changes in database

Modified the submission of votes that voter can cast a vote
without voting any candidate and being tagged as abstained vote

Modified the super admin dashboard, added the impersonation
feature that super admin can view the admin dashboards in a
view mode by impersonation admin scopes

added new analytics for super admin dashboard

modified all of the designs to be matched in the
cvsu theme palette.